more major structural changes
