+++ app/Utility.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App;
+
+class Utility
+{
+    public static function parsePemFromBody($body)
+    {
+        $pem = chunk_split(base64_encode($body), 64, "\n");
+
+        return "-----BEGIN CERTIFICATE-----\n".$pem.'-----END CERTIFICATE-----';
+    }
+
+    public static function base64UrlSafeEncode($input)
+    {
+        //return base64_encode($input);
+        return str_replace('=', '', strtr(base64_encode($input), '+/', '-_'));
+    }
+}
